# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Verdyx API
  version: 1.0.0
  description: |
    Core API gateway routes for Verdyx.
servers:
  - url: http://localhost:8080/api/v1
    description: Local
  - url: https://api-staging.verdyx.com/api/v1
    description: Staging
  - url: https://api.verdyx.com/api/v1
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
...
paths:
...
  /auth/login:
    post:
      summary: Login and get JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Invalid credentials
  /auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
        '400':
          description: Validation failed
...